+++ src/dustapprox/tools/__init__.py
@@ -1,1 +1,6 @@
-""" Convenient tools to generate models and plots. """
+"""Convenient tools to generate models and plots."""
+
+from .downloader import download_file
+
+
+__all__ = ["download_file"]