+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## v0.2.0
 
-- update python supported versions
+- update python supported versions: python >=3.10 <=3.13
 - upgrade packaging to pyproject.toml only
-- typing annotations
+- typing annotations throughout the codebase
+- using [`dust_extinction`](https://dust-extinction.readthedocs.io/en/latest/) package for extinction curves
+- upgraded to [`Pyphot 2.0`](https://github.com/mfouesneau/pyphot/tree/v2.0-dev)
+- Add training scripts to reproduce / add models
+- Add (A0, R0) models
 
 ## v0.1.0